Apply clippy fixes

+++ replay_parser/src/packet_parser/events/game_version.rs
@@ -19,7 +19,7 @@ impl PacketParser for GameVersion {
     fn parse(packet: &Packet, _context: &Context) -> Result<BattleEvent> {
         // First, we skip an le_u32 value which tells us the size of the rest of the packet
         // We don't really need it because we try to match 4 digits in input regardless of size.
-        let (remaining, _) = take(4 as usize)(packet.get_payload())?;
+        let (remaining, _) = take(4_usize)(packet.get_payload())?;
 
         // We could simply the parse the `remaining` as one str but the problem is the delimiter can be
         // different for different replay files so getting that problem out of the way at this stage
